% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planeview.r
\docType{data}
\name{pview.app}
\alias{pview.app}
\title{The name of the plain-view executable. Currently it's the full path to the development
build. Change to just the executable name after it's installed.}
\format{
An object of class \code{character} of length 1.
}
\usage{
pview.app
}
\description{
The name of the plain-view executable. Currently it's the full path to the development
build. Change to just the executable name after it's installed.
}
\keyword{datasets}
