% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planeview.r
\name{pview}
\alias{pview}
\title{Plot vectors}
\usage{
pview(xs, ...)
}
\value{
A string that evaluates to a ggplot of the displayed plot. \emph{Not implemented yet.}
}
\description{
pview may be called as
\enumerate{
\item pview(ys)  A vector of indices will be generated for x.
\item pview(xs, ys1, ys2, ...)
\item pview(list(ys1, ys2, ...))  A vector of indices will be generated for x.
\item pview(list(xs1, xs2, ...), list(ys1, ys2, ...))
}
}
\examples{
xs <- 0:99/10
pview(xs, cos(xs))
pview(xs, cos(xs), sin(xs))

}
