% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planeview.r
\name{pvplot}
\alias{pvplot}
\title{Helper function for pview()}
\usage{
pvplot(xss, yss)
}
\arguments{
\item{xss}{A list of x-axis vectors}

\item{yss}{A list of y-axis vectors}
}
\description{
pvplot() spawns the plotting process and sends the data to it. It's not exported
because it requires a particular format for the data. pview() is a more flexible front
end to this function.
}
